+++ src/Domain/WebRTCToken.php
@@ -11,11 +11,14 @@ use SamuelMwangiW\Africastalking\Saloon\Requests\Voice\CapabilityTokenRequest;
 use SamuelMwangiW\Africastalking\ValueObjects\CapabilityToken;
 use SamuelMwangiW\Africastalking\ValueObjects\PhoneNumber;
 use ReflectionException;
+use InvalidArgumentException;
 
 class WebRTCToken
 {
     public ?string $username = null;
     public ?string $clientName;
+    private int $maxlifeTimeSec = 86400;
+    public int $lifeTimeSec = 86400;
     public ?PhoneNumber $phone;
 
     public function for(?string $name = null): static
@@ -32,6 +35,21 @@ class WebRTCToken
         }
 
         $this->phone = $phoneNo;
+
+        return $this;
+    }
+
+    public function validFor(int $seconds): static
+    {
+        if ($seconds <= 0) {
+            throw new InvalidArgumentException("Negative duration values not allowed");
+        }
+
+        if ($seconds > $this->maxlifeTimeSec) {
+            throw new InvalidArgumentException("The maximum allowed token duration is 24 Hours");
+        }
+
+        $this->lifeTimeSec = $seconds;
 
         return $this;
     }
@@ -83,10 +101,10 @@ class WebRTCToken
 
     public function expire(): string
     {
-        return '86400s';
+        return "{$this->lifeTimeSec}s";
     }
 
-    private function data(): array
+    public function data(): array
     {
         return [
             'phoneNumber' => $this->phone(),