+++ src/Http/Requests/VoiceEventRequest.php
@@ -34,7 +34,7 @@ class VoiceEventRequest extends FormRequest
             'callerCountryCode' => ['nullable', 'string', 'max:4'],
             'callStartTime' => ['nullable', 'string'],
             'recordingUrl' => ['nullable', 'url'],
-            'durationInSeconds  ' => ['nullable', 'int', 'min:0'],
+            'durationInSeconds' => ['nullable', 'int', 'min:0'],
             'currencyCode' => ['nullable', 'string', new Enum(Currency::class)],
             'amount' => ['nullable', 'numeric'],
             'dialDestinationNumber' => ['nullable', 'string'],
@@ -54,12 +54,21 @@ class VoiceEventRequest extends FormRequest
         return 'sessionId';
     }
 
-    public function downloadRecording(string|null $disk = null): void
+    public function downloadRecording(string|null $disk = null, string|null $path = null): void
     {
         if ($this->isEmptyString('recordingUrl')) {
             return;
         }
 
-        DownloadCallRecording::dispatch($this->input('recordingUrl'), $this->id(), $disk);
+        if (0 === $this->integer('durationInSeconds')) {
+            return;
+        }
+
+        DownloadCallRecording::dispatch(
+            $this->input('recordingUrl'),
+            $this->id(),
+            $disk,
+            $path,
+        );
     }
 }