+++ tests/Feature/Requests/VoiceEventRequestTest.php
@@ -2,7 +2,6 @@
 
 declare(strict_types=1);
 
-use Illuminate\Http\Client\RequestException;
 use Illuminate\Support\Facades\Bus;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Http;
@@ -33,6 +32,30 @@ it('Downloading recording dispatches DownloadCallRecording job', function (array
     );
 })->with('voice-event-notification-with-recording');
 
+it('does not download recording when call duration is zero', function (array $notification): void {
+    Bus::fake();
+    Route::post('call', function (VoiceEventRequest $request) {
+        $request->downloadRecording();
+
+        return 'OK';
+    });
+
+    post('/call', $notification);
+    Bus::assertNothingDispatched();
+})->with('voice-event-notification-with-0-duration');
+
+it('does not download recording when recordingUrl is empty', function (array $notification): void {
+    Bus::fake();
+    Route::post('call', function (VoiceEventRequest $request) {
+        $request->downloadRecording();
+
+        return 'OK';
+    });
+
+    post('/call', $notification);
+    Bus::assertNothingDispatched();
+})->with('voice-event-notification-with-empty-recordingUrl');
+
 it('downloads a recording', function (string $url): void {
     Storage::fake();
     Http::fake();
@@ -62,10 +85,25 @@ it('downloads a recording to disk', function (string $url): void {
     fn () => 'https://example.com/Free_Test_Data_100KB_MP3.mp3',
 ]);
 
+it('downloads a recording to a specified path on disk', function (string $url): void {
+    Storage::fake('s3');
+    Http::fake();
+
+    Storage::deleteDirectory('call-recordings');
+
+    DownloadCallRecording::dispatch($url, 'sessionId', 's3', 'path/to/file/example.mp3');
+
+    Storage::disk('s3')
+        ->assertExists('path/to/file')
+        ->assertExists('path/to/file/example.mp3');
+})->with([
+    fn () => 'https://example.com/Free_Test_Data_100KB_MP3.mp3',
+]);
+
 it('can fail to downloads a recording to disk', function (string $url): void {
     Storage::fake('s3');
     Http::fake([
-        '*' => Http::response(null, Response::HTTP_REQUEST_TIMEOUT),
+        '*' => Http::response(null, Response::HTTP_NOT_FOUND),
     ]);
 
     Storage::deleteDirectory('call-recordings');
@@ -73,11 +111,9 @@ it('can fail to downloads a recording to disk', function (string $url): void {
     DownloadCallRecording::dispatch($url, 's3');
 
     Storage::disk('s3')->assertMissing('call-recordings');
-})
-    ->throws(RequestException::class)
-    ->with([
-        fn () => 'https://example.com/Free_Test_Data_100KB_MP3.mp3',
-    ]);
+})->with([
+    fn () => 'https://example.com/Free_Test_Data_100KB_MP3.mp3',
+]);
 
 it('dispatches an event after downloading a recording', function (string $url): void {
     Storage::fake();
@@ -102,7 +138,7 @@ it('dispatches an event after download failed', function (string $url): void {
     Storage::fake();
     Event::fake();
     Http::fake([
-        '*' => Http::response(null, Response::HTTP_REQUEST_TIMEOUT),
+        '*' => Http::response(null, Response::HTTP_NOT_FOUND),
     ]);
 
     Storage::deleteDirectory('call-recordings');
@@ -115,8 +151,6 @@ it('dispatches an event after download failed', function (string $url): void {
             RecordingDownloadFailed $event
         ) => $event->recordingUrl === $url && 'sessionId' === $event->sessionId
     );
-})
-    ->throws(RequestException::class)
-    ->with([
-        'https://example.com/Free_Test_Data_100KB_MP3.mp3',
-    ]);
+})->with([
+    'https://example.com/Free_Test_Data_100KB_MP3.mp3',
+]);